requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
streamlit>=1.20.0
huggingface-hub>=0.14.1
plotly
sqlalchemy >=2.0.0